+++ era-compiler-solidity/src/build_evm/contract.rs
@@ -5,6 +5,7 @@
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
+use std::path::PathBuf;
 
 use crate::solc::combined_json::contract::Contract as CombinedJsonContract;
 use crate::solc::standard_json::output::contract::evm::EVM as StandardJsonOutputContractEVM;
@@ -15,8 +16,8 @@ use crate::solc::standard_json::output::contract::Contract as StandardJsonOutput
 ///
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Contract {
-    /// The contract path.
-    pub path: String,
+    /// The contract name.
+    pub name: era_compiler_common::ContractName,
     /// The auxiliary identifier. Used to identify Yul objects.
     pub identifier: String,
     /// The LLVM deploy code module build.
@@ -32,14 +33,14 @@ impl Contract {
     /// A shortcut constructor.
     ///
     pub fn new(
-        path: String,
+        name: era_compiler_common::ContractName,
         identifier: String,
         deploy_build: era_compiler_llvm_context::EVMBuild,
         runtime_build: era_compiler_llvm_context::EVMBuild,
         metadata_json: serde_json::Value,
     ) -> Self {
         Self {
-            path,
+            name,
             identifier,
             deploy_build,
             runtime_build,
@@ -82,12 +83,21 @@ impl Contract {
     ///
     pub fn write_to_directory(
         self,
-        path: &Path,
+        output_path: &Path,
         _output_assembly: bool,
         output_binary: bool,
         overwrite: bool,
     ) -> anyhow::Result<()> {
-        let (file_name, contract_name) = Self::split_path(self.path.as_str());
+        let file_path = PathBuf::from(self.name.path);
+        let file_name = file_path
+            .file_name()
+            .expect("Always exists")
+            .to_str()
+            .expect("Always valid");
+
+        let mut output_path = output_path.to_owned();
+        output_path.push(file_name);
+        std::fs::create_dir_all(output_path.as_path())?;
 
         if output_binary {
             for (code_type, bytecode) in [
@@ -99,27 +109,25 @@ impl Contract {
             {
                 let output_name = format!(
                     "{}.{}.{}",
-                    contract_name,
+                    self.name.name.as_deref().unwrap_or(file_name),
                     code_type,
                     era_compiler_common::EXTENSION_EVM_BINARY
                 );
-                let mut file_path = path.to_owned();
-                file_path.push(file_name.as_str());
-                std::fs::create_dir_all(file_path.as_path())?;
-                file_path.push(output_name.as_str());
+                let mut output_path = output_path.clone();
+                output_path.push(output_name.as_str());
 
-                if file_path.exists() && !overwrite {
+                if output_path.exists() && !overwrite {
                     anyhow::bail!(
-                        "Refusing to overwrite an existing file {file_path:?} (use --overwrite to force)."
+                        "Refusing to overwrite an existing file {output_path:?} (use --overwrite to force)."
                     );
                 } else {
-                    File::create(&file_path)
+                    File::create(&output_path)
                         .map_err(|error| {
-                            anyhow::anyhow!("File {:?} creating: {}", file_path, error)
+                            anyhow::anyhow!("File {:?} creating: {}", output_path, error)
                         })?
                         .write_all(hex::encode(bytecode.as_slice()).as_bytes())
                         .map_err(|error| {
-                            anyhow::anyhow!("File {:?} writing: {}", file_path, error)
+                            anyhow::anyhow!("File {:?} writing: {}", output_path, error)
                         })?;
                 }
             }
@@ -182,16 +190,5 @@ impl Contract {
             .modify_evm(deploy_bytecode, runtime_bytecode);
 
         Ok(())
-    }
-
-    ///
-    /// Extracts the file and contract names from the full path.
-    ///
-    pub fn split_path(path: &str) -> (String, String) {
-        let path = path.trim().replace(['\\', ':'], "/");
-        let mut path_iterator = path.split('/').rev();
-        let contract_name = path_iterator.next().expect("Always exists");
-        let file_name = path_iterator.next().expect("Always exists");
-        (file_name.to_owned(), contract_name.to_owned())
     }
 }