+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - The compilation pipeline that was not run without output parameters
+- Broken `--output-dir` output paths for non-Solidity contracts
 
 ## [1.5.4] - 2024-09-24